Separada en 6 la colision paleta-pelota, corregidos bugs del power up invisibilidad, eliminados warnings

+++ Pong/console.cpp
@@ -15,7 +15,7 @@ bool pause = 0;
 namespace catPong {
 	void initConsole() {
 		cronometer = (double)clock() / 1000;
-		cronometerFlo = (double)clock() / 1000;
+		cronometerFlo = static_cast<float>(clock()) / 1000;
 		FPS = 120;
 		SetTargetFPS(FPS);
 	}
